requirements.txt
gunicorn==19.9.0
line-bot-sdk==1.12.1
requests==2.22.0
bs4==0.0.1
flask